feat: update AccountCard and AccountsList components to use FinanceData type
